# backend/openapi.yaml
openapi: 3.1.0
info:
  title: itgel API
  version: 1.0.0
  description: |
    Захиалгын дэлгүүрийн backend API.

    - Бүх мөнгөн дүн бүхэл тоо (₮).
    - Бүх огноо ISO 8601, `Asia/Ulaanbaatar` (UTC+8) цагаар бодогдоно.
    - Хэрэглэгчийн API `costPrice`-г хэзээ ч буцаахгүй.
servers:
  - url: http://localhost:4000
    description: Dev
...
security: []
paths:
...
  /api/me/orders:
    get:
      tags: [public]
      security: [{ bearerAuth: [] }]
      summary: Өөрийн захиалгын түүх (timeline-тай)
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Захиалгууд ба нийлбэрүүд
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/PublicOrder' }
                  meta:
                    allOf:
                      - $ref: '#/components/schemas/PageMeta'
                      - type: object
                        properties:
                          totalSpent: { type: integer }
                          activeCount: { type: integer }
  /api/orders:
    post:
      tags: [public]
      summary: Захиалга үүсгэх
      description: |
        IP-ээр 10 минутад 10 удаа хязгаартай.
        Нэг checkout үйлдэлд `Idempotency-Key` илгээнэ. Ижил хэрэглэгч+key+payload
        дахин ирвэл өмнөх захиалгын хариу (mixed-cart `splitOrders` оролцуулаад) буцна.
        Ижил key өөр payload-тай бол 409. Шинэ худалдан авалтад шинэ key хэрэгтэй.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                name: { type: string }
                note: { type: string }
                leasing: { type: boolean }
                idempotencyKey: { type: string, minLength: 8, maxLength: 128 }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, qty]
                    properties:
                      productId: { type: string }
                      qty: { type: integer, minimum: 1 }
                      selections:
                        type: object
                        additionalProperties: { type: string }
                      size: { type: string }
                      color: { type: string }
      responses:
        '201':
          description: Үүссэн захиалга эсвэл ижил checkout-ийн хадгалсан хариу
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      code: { type: string, example: PH-K7M2XQ }
                      status: { $ref: '#/components/schemas/OrderStatus' }
                      subtotal: { type: integer }
                      dueAmount: { type: integer }
                      splitOrders:
                        type: array
                        items:
                          type: object
                          properties:
                            code: { type: string }
                            payeeKind: { type: string }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/orders/{code}:
    get:
      tags: [public]
      summary: Захиалгын хяналт (timeline-тай)
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Захиалга
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/PublicOrder' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/admin/orders:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Захиалгын жагсаалт
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
        - { name: q, in: query, schema: { type: string }, description: 'Код, утас, нэрээр' }
        - { name: batch, in: query, schema: { type: string } }
        - { name: fulfilment, in: query, schema: { type: string, enum: [PICKUP, DELIVERY] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200': { description: Жагсаалт }
  /api/admin/orders/{id}:
    get:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Захиалгын дэлгэрэнгүй
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Захиалга }
  /api/admin/orders/{id}/status:
    patch:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Төлөв солих
      description: |
        Зөвшөөрөгдөх шилжилт: NEW→CONFIRMED→IN_BATCH→IN_TRANSIT→ARRIVED→HANDED_OVER.
        CANCELLED руу HANDED_OVER-с бусад бүх төлвөөс. Буруу шилжилтэд 409.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/OrderStatus' }
                reason: { type: string }
      responses:
        '200': { description: Шинэчлэгдсэн захиалга }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/admin/batches/{id}/orders:
    post:
      tags: [admin]
      security: [{ bearerAuth: [] }]
      summary: Захиалга нэмэх, хасах
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add: { type: array, items: { type: string } }
                remove: { type: array, items: { type: string } }
      responses:
        '200': { description: Нэмэгдсэн/хасагдсан тоо }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Манай JWT (`/api/auth/verify` эсвэл `/api/admin/auth/login`) эсвэл
        Supabase Auth-ийн access token (JWKS-ээр шалгагдана).
  responses:
...
    NotFound:
      description: Олдсонгүй
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Бизнес дүрэм зөрчсөн
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        pages: { type: integer }
...
    OrderStatus:
      type: string
      enum: [NEW, CONFIRMED, IN_BATCH, IN_TRANSIT, ARRIVED, HANDED_OVER, CANCELLED]
...
    TimelineStep:
      type: object
      properties:
        key:
          type: string
          enum: [placed, confirmed, sent_to_supplier, in_transit, arrived, handed_over, cancelled]
        label: { type: string }
        status: { type: string, enum: [done, current, pending] }
        at: { type: string, format: date-time, nullable: true }
        estimatedAt: { type: string, format: date-time, nullable: true }
    PublicOrder:
      type: object
      properties:
        code: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        statusLabel: { type: string }
        subtotal: { type: integer }
        paidAmount: { type: integer }
        refundedAmount: { type: integer }
        dueAmount: { type: integer }
        deliveryFee: { type: integer }
        paymentState:
          type: string
          enum: [UNPAID, PARTIAL, PAID, OVERPAID, REFUNDED]
        fulfilment: { type: string, enum: [PICKUP, DELIVERY], nullable: true }
        canChooseFulfilment: { type: boolean }
        items:
          type: array
          items:
            type: object
            properties:
              productId: { type: string }
              name: { type: string }
              size: { type: string, nullable: true }
              color: { type: string, nullable: true }
              qty: { type: integer }
              unitPrice: { type: integer }
              total: { type: integer }
        timeline:
          type: array
          items: { $ref: '#/components/schemas/TimelineStep' }
...